+++ src/main/java/xyz/geik/glib/economy/GrinGotts.java
@@ -36,7 +36,7 @@ class GrinGotts implements Economy {
      * @return price
      */
     @Override
-    public double withdrawPlayer(OfflinePlayer player, long price) {
+    public double withdrawPlayer(OfflinePlayer player, double price) {
         if (this.economy != null) {
             this.economy.player(player.getUniqueId()).withdraw(price);
         }
