+++ src/main/java/xyz/geik/glib/economy/RoyaleEconomy.java
@@ -35,7 +35,7 @@ class RoyaleEconomy implements Economy {
      * @return price
      */
     @Override
-    public double withdrawPlayer(OfflinePlayer player, long price) {
+    public double withdrawPlayer(OfflinePlayer player, double price) {
         if (this.economy != null) {
             this.economy.balance.removeBankBalance(player.getUniqueId() + "", price);
         }
